initial commit
Takes a list of directories to walk and searches all YAML and JSON files
in all directories / their subdirectories for any key with the name "image" and
saves the value to which said key maps. This allows for easily seeing what container
images are being used by all YAML and JSON files in any directory.

Also takes an option to do some validation of the image names, which checks for the following:
* Whether an exact version of the image is specified (e.g., not "latest")
* Checks that the image is from a specified container image hosting service (e.g., quay.io, DockerHub, etc.).

Signed-off by: Ian Vernon <ianvernon92@gmail.com>